remove directory for balsamic specific rules
