+++ cg/services/orders/validation/rules/case_sample/utils.py
@@ -2,6 +2,7 @@ import re
 from collections import Counter
 
 from cg.constants.constants import StatusOptions
+from cg.constants.sequencing import SeqLibraryPrepCategory
 from cg.constants.subject import Sex
 from cg.models.orders.sample_base import ContainerEnum, SexEnum
 from cg.services.orders.validation.errors.case_errors import RepeatedCaseNameError
@@ -33,6 +34,8 @@ from cg.services.orders.validation.rules.utils import (
     is_sample_on_plate,
     is_volume_within_allowed_interval,
 )
+from cg.services.orders.validation.workflows.balsamic.models.sample import BalsamicSample
+from cg.store.models import Application
 from cg.store.store import Store
 
 
@@ -282,4 +285,14 @@ def is_buffer_missing(sample: SampleInCase) -> bool:
     return bool(
         sample.application.startswith(tuple(applications_requiring_buffer))
         and not sample.elution_buffer
+    )
+
+
+def is_sample_missing_capture_kit(sample: BalsamicSample, store: Store) -> bool:
+    """Returns whether a TGS sample has an application and is missing a capture kit."""
+    application: Application | None = store.get_application_by_tag(sample.application)
+    return (
+        application
+        and application.prep_category == SeqLibraryPrepCategory.TARGETED_GENOME_SEQUENCING
+        and not sample.capture_kit
     )
