+++ cg/services/orders/validation/workflows/balsamic/validation_rules.py
@@ -11,6 +11,7 @@ from cg.services.orders.validation.rules.case_sample.rules import (
     validate_application_not_archived,
     validate_buffer_required,
     validate_buffer_skip_rc_condition,
+    validate_capture_kit_panel_requirement,
     validate_concentration_interval_if_skip_rc,
     validate_concentration_required_if_skip_rc,
     validate_container_name_required,
@@ -26,9 +27,6 @@ from cg.services.orders.validation.rules.case_sample.rules import (
     validate_well_position_format,
     validate_well_positions_required,
     validate_wells_contain_at_most_one_sample,
-)
-from cg.services.orders.validation.workflows.balsamic.rules import (
-    validate_capture_kit_panel_requirement,
 )
 
 BALSAMIC_CASE_RULES: list[callable] = [
